+++ src/Enum/WepKeyStorage.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum WepKeyStorage: string
+{
+    case Hex = 'H';
+    case StringValue = 'S';
+}